Be more robust about checking VERBOSE value
Leaving `VERBOSE` as undefined causes these intermittent errors:

```
/usr/local/sbin/hydrate-ips.sh: line 17: [: -eq: unary operator expected
```

This is because the `-eq` operator expects a number, which is hard to evaluate if `VERBOSE` is undefined. We could fix this by pre-defining `export VERBOSE=0` but it feels like the wrong approach; ultimately we want to know if `VERBOSE` is set, not what it is set to. Therefore we'll use the `-n` operator instead, which evaluates true for a non-null value.