Only reload nginx if necessary
Right now we're just reloading nginx at the end of the script whether or not we need to. This is in case the content of `/etc/nginx/real_ip.conf` has changed. Rather than guessing, there's an easy way to find out: we'll md5 the config file before it's overwritten (or use a pretend hash if it doesn't exist), and then again afterwards, and compare the results. If they don't match, reload nginx.