Use new CF URL for IPs; fix regex
A few nights ago, the URL for Cloudflare's IPv4 addresses was `https://www.cloudflare.com/ips-v4`; as of tonight it's `https://www.cloudflare.com/ips-v4/` -- not the trailing slash. The IPv6 URL has seen the same change.

In addition to the URL change, the output has changed a bit. Whereas it used to be completely plaintext, a couple HTML tags have now snuck in...

```
$ curl -Ss https://www.cloudflare.com/ips-v6/
<pre style="word-wrap: break-word; white-space: pre-wrap;">2400:cb00::/32
2405:8100::/32
2405:b500::/32
2606:4700::/32
2803:f800::/32
2c0f:f248::/32
2a06:98c0::/29
</pre>
```

To address this, the grep is applied to the whole list in one fell swoop with the `-o` option so that only exact matches are outputted, and not matching _lines_ (which would not be helpful for the first line!) This should be safer (no accidental skunky lines) and is a bit simpler (no applying the regex line-by-line-by-line).